RPI sensor readout/requirements.txt
cachetools==3.0.0
certifi==2018.11.29
chardet==3.0.4
google-api-core==1.7.0
google-auth==1.6.2
google-cloud-pubsub==0.39.1
googleapis-common-protos==1.5.5
gpiozero==1.4.1
grpc-google-iam-v1==0.11.4
grpcio==1.17.1
idna==2.8
micropyGPS==1.0
pigpio==1.42
protobuf==3.6.1
pyasn1==0.4.5
pyasn1-modules==0.2.3
pynmea2==1.12.0
pyserial==3.4
python-dotenv==0.10.1
pytz==2018.9
requests==2.21.0
RPi.GPIO==0.6.5
RPIO==0.10.0
rsa==4.0
six==1.12.0
smbus2==0.2.3
urllib3==1.24.1